% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.effort.R
\name{allocate.effort}
\alias{allocate.effort}
\title{Allocate stations across strata}
\usage{
allocate.effort(
  survey,
  channel = NA,
  schema,
  pwrd,
  points,
  stations.available,
  number.of.tows = 741,
  output.file
)
}
\arguments{
\item{survey}{character string. Defaults to "AI" for Aleutian Islands}

\item{channel}{open connection object. Created from RODBC::odbcConnect()}

\item{schema}{character string. SQL schema name. Defaults to AIGOA_WORK_DATA}

\item{pwrd}{character string. SQL password}

\item{points}{dataframe of potential station locations queried from SQL,
called in AIGOASurveyPlanning::goa.planning().}

\item{stations.available}{integer. Vector of available stations in each
stratum.}

\item{number.of.tows}{integer. Number of total stations.}

\item{output.file}{character string. Output directory.}
}
\value{

}
\description{
Allocates stations across strata based on: stratum size, historical standard
devation of cpue, ex vessel price, and abundance of a preselected subset of
species, and the time it takes to complete tows. (Reference a tech memo for
how exactly Neyman Allocations are handled?).

Updates of code from Ned:
Initial changes are to reformat code into my preferred configuration
Added ", believeNRows = FALSE" to odbc calls for version 3.0.1 of R

Added package XLConnect to provide functionality to to read Excel
workbooks under version 3.0.1 of R in 64-bit environment

Created 12/18/13
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
