% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dd_to_mins.R
\name{convert_dd_to_mins}
\alias{convert_dd_to_mins}
\title{Converts Latitude or Longitude from Decimal Degrees to Minutes?}
\usage{
convert_dd_to_mins(x, digits = 2)
}
\arguments{
\item{x}{vector of latitudes or longitudes to convert.}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\description{
Station lat/lon are converted from decimal degrees to
minutes for the skipper paper station logs
}
\author{
Zack Oyafuso \email{zack.oyafuso@noaa.gov}
}
