% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{D_gct}
\alias{D_gct}
\title{Name_of_the_data3}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}
}
\usage{
D_gct
}
\description{
Description_of_the_Data
}
\keyword{datasets}
