% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{depth_mods_2023}
\alias{depth_mods_2023}
\title{Name_of_the_data}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}
}
\usage{
depth_mods_2023
}
\description{
Description_of_the_Data
}
\keyword{datasets}
