% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.planning.data.R
\name{get.planning.data}
\alias{get.planning.data}
\title{Stratum-level catch stats}
\usage{
get.planning.data(
  channel = NA,
  survey = "AI",
  schema = "AIGOA_WORK_DATA",
  pwrd = pwrd
)
}
\arguments{
\item{channel}{open connection object. Created from RODBC::odbcConnect()}

\item{survey}{character string. Defaults to "AI" for Aleutian Islands}

\item{schema}{character string. SQL schema name. Defaults to AIGOA_WORK_DATA}

\item{pwrd}{character string. SQL password}
}
\value{
A dataframe containing:\tabular{ll}{
   \code{year} \tab integer. Year of survey. \cr \tab \cr
   \code{stratum} \tab integer. Stratum id. \cr \tab \cr
   \code{species_code} \tab integer. Species code. \cr \tab \cr
   \code{haul_count} \tab integer. Total number of hauls. \cr \tab \cr
   \code{cpue} \tab numeric. Average CPUE. \cr \tab \cr
   \code{cpue_std} \tab numeric. Standard deviation of CPUE. \cr \tab \cr
}

These values inform the allocation of stations (effort) to survey strata.
}
\description{
Planning data were generated from a schema-dependent stored procedure
(STATION_PLANNING) in Oracle. Average and standard deviation of CPUE hauls
for a subset of preselected species were calculated for each year and
stratum id.
Created 06/19/2019
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
