% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa.planning.R
\name{goa.planning}
\alias{goa.planning}
\title{Main wrapper function for planning GAP STRS survey in the AI-GOA}
\usage{
goa.planning(selection.type = "random", min.size = 5)
}
\description{
Modifications as of 6/19/2019:
Encapsulated creation of the AI_STATIONS list of candidate stations and
creating the PLANNING_DATA table in R functions called from goa.planning()
and allocate.effort, respectively. This entailed intake of uid, pwd
parameters and passing those along with with the survey parameter to those
functions. Also changed data constraints for AI stations to include good
Ocean Hope stations from 1991 that have yet to be replicated since and
conditioned the general rule of using "previously successful tows" with
consideration of trawlability field populated in AI.AIGRID_GIS from
electronic station logs maintained on the vessel. The routine as written in
this project is functional only for the Aleutian Islands. Gulf of Alaska
station allocation originates from old SPLUS code curated by Paul von Szalay.

Modifications as of 1/20/2016
Added a csv output in addition to the xlsx workbook and identified it with
number.of.tows

Modification as of 12/19/2013
Added readline and choose.dir functions to accept parameter inputs. Changed
write.table for reporting allocation table to writing an XLSX file

Modifications as of 12/18/2013
Initially just changing indentation and line breaks to get preferred code
form and added ", believeNRows = FALSE" to sqlQuery's to deal with version
3.0.1 of R.
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
