% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_allocate_stations.R
\name{goa_allocate_stations}
\alias{goa_allocate_stations}
\title{Gulf of Alaska Bottom Trawl Survey: Allocate stations across strata}
\usage{
goa_allocate_stations(
  n = 550,
  species = c("arrowtooth flounder", "Pacific cod", "walleye pollock", "rex sole",
    "flathead sole", "Pacific halibut", "southern rock sole", "northern rock sole",
    "Pacific ocean perch", "silvergray rockfish", "northern rockfish", "dusky rockfish",
    "BS and RE rockfishes", "Dover sole", "shortspine thornyhead")
)
}
\arguments{
\item{n}{integer.Total number of stations}

\item{species}{character vector. See details for full list of species}
}
\value{
A named vector of stations across strata
}
\description{
To be filled in
}
\author{
Zack Oyafuso \email{zack.oyafuso@noaa.gov}
}
