% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goa_stations}
\alias{goa_stations}
\title{Gulf of Alaska (GOA) Bottom Trawl Survey Stations}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{GRIDID}{Position of the 5-km grid cell }
  \item{NMFS_AREA}{NMFS statistical area name.}
  \item{REP_AREA}{Numeric code for NMFS statistical areas.}
  \item{STRATUM}{Numeric code for GOA stratum.}
  \item{TRAWLABLE}{Character string denoting trawlability status. Valid
  values: 'Y' for trawlable, 'N' for untrawlable, and 'UNK' for unknown.}
  \item{AREA_KM2}{Numeric, station area in square kilometers}
}
}
\usage{
goa_stations
}
\description{
Includes trawlability information from 2023
}
\keyword{datasets}
