% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goa_stratum_boundaries}
\alias{goa_stratum_boundaries}
\title{2025 Gulf of Alaska Stratum Boundaries}
\format{
A data frame with fields:
\describe{
  \item{NMFS_AREA}{NMFS statistical area name.}
  \item{REP_AREA}{Numeric code for NMFS statistical areas.}
  \item{STRATUM}{Numeric code for GOA stratum.}
  \item{DEPTH_MIN_M}{Numeric. Minimum boundary of the stratum in meters.}
  \item{DEPTH_MAX_M}{Numeric. Maximum boundary of the stratum in meters.}
  \item{AREA_KM2}{Numeric. Total area of the stratum in square kilometers.}
  \item{PERIM_KM}{Numeric. Total perimeter of the stratum in kilometers.}
  \item{USED}{Boolean. Is the stratum currently used in the GOA survey footprint? This field is relevant when considering the GOA survey footprint (up to 700 m or 1000 m?).}
}
}
\usage{
goa_stratum_boundaries
}
\description{
data frame loop up table of depth stratum boundaries by NMFS management area used for the Gulf of Alaska bottom trawl survey design (2025-on).
}
\keyword{datasets}
