% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_running_headers.R
\name{insert_running_headers}
\alias{insert_running_headers}
\title{Create a function to insert a header row at a set interval}
\usage{
insert_running_headers(
  df,
  n = 13,
  title_column_idx = 4,
  title = "Alaska Provider FPC Station Log 202501"
)
}
\arguments{
\item{df}{dataframe of the station log}

\item{n}{interger. At which inverval should a page break be made? Default
is 13, means that title and header headers are inserted every 13 records.}

\item{title_column_idx}{integer. Column index where the title should be placed}

\item{title}{string. Running title of the log, e.g.,
"Alaska Provider FPC Station Log 202501"}
}
\description{
Inserts running breaks for header and title rows during
the production of the paper FPC and Skipper Station Log Sheets
}
\author{
Zack Oyafuso \email{zack.oyafuso@noaa.gov}
}
