% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{optim_df}
\alias{optim_df}
\title{Name_of_the_data4}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}
}
\usage{
optim_df
}
\description{
Description_of_the_Data
}
\keyword{datasets}
