% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.gridpoints.R
\name{pick.gridpoints}
\alias{pick.gridpoints}
\title{Randomly sample stations}
\usage{
pick.gridpoints(allocations, user.name.password, survey, points, min.size = 5)
}
\arguments{
\item{allocations}{integer. Named vector of station allocations across strata..}

\item{survey}{character string. Defaults to "AI" for Aleutian Islands}

\item{points}{dataframe of potential station locations queried from SQL,
called in AIGOASurveyPlanning::goa.planning().}
}
\description{
Based on a particular allocation of stations across strata, randomly sample
station locations.
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
