% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.allocations.by.stratum.R
\name{plot.allocations.by.stratum}
\alias{plot.allocations.by.stratum}
\title{Wrapper function for plotting allocation plots by stratum}
\usage{
\method{plot}{allocations.by.stratum}(answer.years, number.tows, strata, channel, survey, output.file)
}
\arguments{
\item{answer.years}{numeric array of dimensions (number of strata, number of
species, number of years) of the weighted optimal allocation (see reference...)}

\item{number.tows}{numeric. Vector output from
AIGOASurveyPlanning::allocate.effort}

\item{strata}{Dataframe. Output from AIGOASurveyPlanning::allocate.effort}

\item{channel}{open connection object. Created from RODBC::odbcConnect()}

\item{survey}{character string. Defaults to "AI" for Aleutian Islands}

\item{output.file}{character string. Output directory.}
}
\value{
AllocationPlotsByStratum.pdf saved in the output.file directory
}
\description{
Creates AllocationPlotsByStratum.pdf, which includes output from
AIGOASurveyPlanning::plot.strata.stations as well as other output (finish!)

Created 12/18/13

Notes from Ned: added ", believeNRows = FALSE" to deal with version 3.0.1 of
R in the 64-bit environment
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
