% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_simulation_functions.R
\name{sample_srs}
\alias{sample_srs}
\title{Simulate a simple random design Gulf of Alaska survey}
\usage{
sample_srs(d_gc = NULL, sample_size = 550)
}
\arguments{
\item{d_gc}{matrix with dimensions n_cells x n_spp where n_cells is the
total number of cells and n_spp is the total number of species}

\item{sample_size}{integer, total number of random samples}
}
\value{
dataframe of mean, sd
}
\description{
Internal function used in simulate_goa_strs()
}
\author{
Zack Oyafuso \email{zack.oyafuso@noaa.gov}
}
