% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_simulation_functions.R
\name{simulate_goa_strs}
\alias{simulate_goa_strs}
\title{Simulate a stratified random design Gulf of Alaska survey}
\usage{
simulate_goa_strs(d_gct = NULL, grid_df = NULL, stn_allocation = goa_stn_all)
}
\arguments{
\item{d_gct}{array with dimensions n_cells x n_spp, n_years where n_cells
is the total number of cells and n_spp is the total number of
species, and n_years is the total number of survey years}

\item{grid_df}{data.frame with n_cells records, contains at least a column
called "STRATUM" that has the stratum information}

\item{stn_allocation}{data.frame with n_strata records, where n_strata
is the total number of survey strata, contains
columns "stratum" and "n" for the station allocation.}
}
\value{
named list of stratum_info, stratified random means and cvs by
         species and year.
}
\description{
Simulate a stratified random design Gulf of Alaska survey
}
\author{
Zack Oyafuso \email{zack.oyafuso@noaa.gov}
}
