% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.allocation.R
\name{vessel.allocation}
\alias{vessel.allocation}
\title{Allocate stations across vessels}
\usage{
vessel.allocation(stations, strata, survey)
}
\arguments{
\item{stations}{dataframe of sampled stations created interanlly in
AIGOASurveyPlanning::goa.planning()}

\item{strata}{dataframe of strata characteristics created by
AIGOASurveyPlanning::allocate.effort()}

\item{survey}{character string. Defaults to "AI" for Aleutian Islands}
}
\description{
Allocate stations across vessels
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
