% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.effort.R
\name{allocate.effort}
\alias{allocate.effort}
\title{Allocate stations across strata}
\usage{
allocate.effort(
  channel = NULL,
  species_weightings = NULL,
  stratum_stats = NULL,
  n_total = 400,
  min_n_stratum = 2
)
}
\arguments{
\item{channel}{open connection object. Created from gapindex::get_connected()}

\item{species_weightings}{dataframe of species codes and weightings}

\item{stratum_stats}{dataframe of stratum means and sds for each species}

\item{n_total}{integer. Number of total stations}

\item{min_n_stratum}{integer. Minimum number of stations in a stratum}
}
\value{
A named vector of allocated stations across strata
}
\description{
Allocates stations across strata based on: stratum size, historical standard
deviation of cpue, ex vessel price, and abundance of a preselected subset of
species (reference a tech memo for how exactly Neyman allocations are handled?).

Created 12/18/13
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
