% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{D_gct}
\alias{D_gct}
\title{Predicted densities: Gulf of Alaska Groundfish, 1996-present}
\format{
A 3-dimensional array of predicted densities, units kg/km2
\describe{
  \item{grid cell}{VAST interpolation grid cell, denoted by subscript g}
  \item{taxon}{species/complex, denoted by subscript c}
  \item{time}{year, denoted by subscript t}
  }
}
\usage{
D_gct
}
\description{
Predicted densities (kg/km2) from VAST for 15 species/complexes from
1996-present (excluding 2001).
}
\keyword{datasets}
