% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ai.stations.R
\name{get.ai.stations}
\alias{get.ai.stations}
\title{Query available Aleutian Islands stations}
\usage{
get.ai.stations(
  channel = NA,
  schema = "AIGOA_WORK_DATA",
  pwrd = NA,
  survey = "AI"
)
}
\arguments{
\item{channel}{open connection object. Created from RODBC::odbcConnect()}

\item{schema}{character string. SQL schema name. Defaults to AIGOA_WORK_DATA}

\item{pwrd}{character string. SQL password}

\item{survey}{character string. Defaults to "AI" for Aleutian Islands}
}
\value{
A dataframe of candidate stations (field stationid) for Aleutian
 Islands bottom trawl surveys which includes the station centroid (fields
 longitude & latitude in decimal degrees) and stratum id.
}
\description{
This script is a translation of the PL/SQL procedure stored on the AI schema
in Oracle. It utilizes the functionality of creating/dropping tables in
Oracle using sqlSave in place of the Global Temporary table strategy
utilized in the PL/SQL procedure. This function also updates the selection
of AI candidate stations to incorporate designations of trawlability stored
in the AI.AIGRID_GIS table and originating with the electronic station logs
kept aboard the vessel by the FPC.

Date last modified: 06/19/2019
}
\author{
Ned Laman \email{ned.laman@noaa.gov}
}
