% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{optim_df}
\alias{optim_df}
\title{VAST interpolation grid}
\format{
A data frame with fields:
\describe{
  \item{ID}{Integer. Unique cell id.}
  \item{Area_km2}{Numeric. total area of the cell in square kilometers.}
  \item{Depth_m}{Numeric. Depth of the cell in meters}
  \item{Eastings}{Numeric. Eastings in meters using NAD83 / Alaska Albers (EPSG:3338)}
  \item{Northings}{Numeric. Northings in meters using NAD83 / Alaska Albers (EPSG:3338)}
  \item{NMFS_AREA}{NMFS statistical area name.}
  \item{STRATUM}{Numeric code for GOA stratum.}
}
}
\usage{
optim_df
}
\description{
data frame of VAST interpolation grid cells used in the optimization of
station effort across Gulf of Alaska bottom trawl survey strata (2025-on).
}
\keyword{datasets}
